Add Untrack APK task
This commit adds untrackApk task for the plugin to allow user to untrack
app from another Play Store channel when they are using 2 different
channels to deploy their app.

```
gradle untrackApkRelease
```

This commit adds `untrack` and `untrackFormat` properties to the
extension to allow further customisation on which version of the app to
be disabled from Play Store.

```groovy
play {
  // untrack from alpha any active versions between 300 and 399
  untrack = 'alpha'
  untrackFormat = '^3[0-9]{2}\$'
}
```